Rename attribute tests and drop ClassWithAttributesAllow
... which was a copy of ClassWithAttributes but with wrong comments. Instead of making the comments right, I've dropped it and had to split and rename some tests.
